INT-491 Replace text log viewer with xterm.js terminal
Worker containers produce raw VT100/ANSI terminal output that was garbled
in the text-based LogViewer. New TerminalLogViewer uses xterm.js to properly
render escape sequences, colors, and cursor positioning.
