+++ apps/web/package.json
@@ -24,6 +24,8 @@
     "@types/js-yaml": "^4.0.9",
     "@uiw/react-markdown-preview": "^5.1.5",
     "@uiw/react-md-editor": "^4.0.11",
+    "@xterm/addon-fit": "^0.11.0",
+    "@xterm/xterm": "^6.0.0",
     "firebase": "^11.9.0",
     "highlight.js": "^11.11.1",
     "js-yaml": "^4.1.1",
