+++ apps/web/src/pages/CodeTaskDetailPage.tsx
@@ -1,4 +1,4 @@
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { memo, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import {
   AlertCircle,
@@ -12,58 +12,13 @@ import {
   Link2,
   RotateCcw,
   StopCircle,
-  Terminal,
   XCircle,
 } from 'lucide-react';
-import {
-  collection,
-  onSnapshot,
-  orderBy,
-  query,
-  type Unsubscribe,
-} from 'firebase/firestore';
 import { Button, Card, Layout } from '@/components';
-import { useAuth } from '@/context';
+import { TerminalLogViewer } from '@/components/TerminalLogViewer.js';
 import { useCodeTask } from '@/hooks';
-import {
-  getFirestoreClient,
-  authenticateFirebase,
-  isFirebaseAuthenticated,
-  initializeFirebase,
-} from '@/services/firebase';
 import { formatDateTime, formatRelative } from '@/utils/dateFormat';
 import type { CodeTask, CodeTaskStatus } from '@/types';
-
-interface LogEntry {
-  id: string;
-  timestamp: string;
-  level: 'info' | 'warn' | 'error' | 'debug';
-  message: string;
-  tool?: string;
-}
-
-// Helper functions moved outside components to avoid recreation on each render
-const formatLogTime = (timestamp: string): string => {
-  const date = new Date(timestamp);
-  return date.toLocaleTimeString('en-US', {
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-  });
-};
-
-const getLevelColor = (level: LogEntry['level']): string => {
-  switch (level) {
-    case 'error':
-      return 'text-red-400';
-    case 'warn':
-      return 'text-amber-400';
-    case 'debug':
-      return 'text-slate-500';
-    default:
-      return 'text-green-400';
-  }
-};
 
 interface StatusStyle {
   bg: string;
@@ -241,7 +196,7 @@ export function CodeTaskDetailPage(): React.JSX.Element {
 
       {task.error !== undefined ? <TaskErrorCard task={task} /> : null}
 
-      <LogViewer taskId={task.id} isActive={isRunning} />
+      <TerminalLogViewer taskId={task.id} isActive={isRunning} />
     </Layout>
   );
 }
@@ -538,254 +493,3 @@ const TaskErrorCard = memo(function TaskErrorCard({ task }: TaskErrorCardProps):
   return prevProps.task.error === nextProps.task.error;
 });
 
-// Memoized LogViewer - only re-renders when taskId or isActive changes
-interface LogViewerProps {
-  taskId: string;
-  isActive: boolean;
-}
-
-const LogViewer = memo(function LogViewer({ taskId, isActive }: LogViewerProps): React.JSX.Element {
-  const { getAccessToken } = useAuth();
-  const [logs, setLogs] = useState<LogEntry[]>([]);
-  const [logsLoading, setLogsLoading] = useState(true);
-  const [logsError, setLogsError] = useState<string | null>(null);
-  const [autoScroll, setAutoScroll] = useState(true);
-  const [logsHeight, setLogsHeight] = useState(384); // max-h-96 = 384px
-  const [isResizing, setIsResizing] = useState(false);
-  const [copiedLogId, setCopiedLogId] = useState<string | null>(null);
-
-  const logsEndRef = useRef<HTMLDivElement>(null);
-  const logsContainerRef = useRef<HTMLDivElement>(null);
-  const resizeHandleRef = useRef<HTMLDivElement>(null);
-  const unsubscribeRef = useRef<Unsubscribe | null>(null);
-  const firebaseAuthenticatedRef = useRef(false);
-  const isMountedRef = useRef(true);
-  // Use ref for autoScroll to avoid dependency cycles in useEffect
-  const autoScrollRef = useRef(autoScroll);
-
-  // Keep the ref in sync with state
-  useEffect(() => {
-    autoScrollRef.current = autoScroll;
-  }, [autoScroll]);
-
-  const scrollToBottom = useCallback((): void => {
-    if (autoScrollRef.current && logsContainerRef.current !== null) {
-      logsContainerRef.current.scrollTop = logsContainerRef.current.scrollHeight;
-    }
-  }, []); // No dependencies - uses ref instead
-
-  const handleMouseDown = useCallback((e: React.MouseEvent) => {
-    e.preventDefault();
-    setIsResizing(true);
-
-    const startY = e.clientY;
-    const startHeight = logsHeight;
-
-    const handleMouseMove = (moveEvent: MouseEvent): void => {
-      const deltaY = startY - moveEvent.clientY;
-      const newHeight = Math.max(200, Math.min(800, startHeight + deltaY));
-      setLogsHeight(newHeight);
-    };
-
-    const handleMouseUp = (): void => {
-      setIsResizing(false);
-      document.removeEventListener('mousemove', handleMouseMove);
-      document.removeEventListener('mouseup', handleMouseUp);
-    };
-
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp);
-  }, [logsHeight]);
-
-  const copyLogEntry = useCallback((logMessage: string, logId: string) => {
-    void navigator.clipboard.writeText(logMessage);
-    setCopiedLogId(logId);
-    setTimeout(() => {
-      setCopiedLogId(null);
-    }, 2000);
-  }, []);
-
-  useEffect(() => {
-    isMountedRef.current = true;
-    return (): void => {
-      isMountedRef.current = false;
-    };
-  }, []);
-
-  useEffect(() => {
-    const setupListener = async (): Promise<void> => {
-      try {
-        if (!firebaseAuthenticatedRef.current || !isFirebaseAuthenticated()) {
-          initializeFirebase();
-          const token = await getAccessToken();
-          await authenticateFirebase(token);
-          firebaseAuthenticatedRef.current = true;
-        }
-
-        const db = getFirestoreClient();
-        const logsRef = collection(db, 'code_tasks', taskId, 'logs');
-        const logsQuery = query(logsRef, orderBy('timestamp', 'asc'));
-
-        unsubscribeRef.current = onSnapshot(
-          logsQuery,
-          (snapshot) => {
-            if (!isMountedRef.current) return;
-
-            const newLogs: LogEntry[] = [];
-            snapshot.forEach((doc) => {
-              const data = doc.data();
-              const toolValue = data['tool'] as string | undefined;
-              const levelValue = data['level'] as string | undefined;
-              const level: LogEntry['level'] =
-                levelValue === 'info' ||
-                levelValue === 'warn' ||
-                levelValue === 'error' ||
-                levelValue === 'debug'
-                  ? levelValue
-                  : 'info';
-              // Firestore Timestamp has toDate() method, convert to ISO string
-              const timestampField = data['timestamp'];
-              const timestamp =
-                timestampField !== null &&
-                typeof timestampField === 'object' &&
-                'toDate' in timestampField &&
-                typeof timestampField.toDate === 'function'
-                  ? (timestampField.toDate() as Date).toISOString()
-                  : String(timestampField);
-              const entry: LogEntry = {
-                id: doc.id,
-                timestamp,
-                level,
-                message: data['content'] as string,
-              };
-              if (toolValue !== undefined) {
-                entry.tool = toolValue;
-              }
-              newLogs.push(entry);
-            });
-            setLogs(newLogs);
-            setLogsLoading(false);
-            setTimeout(scrollToBottom, 100);
-          },
-          (err) => {
-            if (isMountedRef.current) {
-              setLogsError(err.message);
-              setLogsLoading(false);
-            }
-          }
-        );
-      } catch (err) {
-        if (isMountedRef.current) {
-          setLogsError(err instanceof Error ? err.message : 'Failed to load logs');
-          setLogsLoading(false);
-        }
-      }
-    };
-
-    void setupListener();
-
-    return (): void => {
-      if (unsubscribeRef.current !== null) {
-        unsubscribeRef.current();
-        unsubscribeRef.current = null;
-      }
-    };
-  }, [taskId, getAccessToken, scrollToBottom]); // scrollToBottom is now stable (empty deps)
-
-  // Memoize logs rendering to avoid recalculating on every parent re-render
-  const logsElements = useMemo(() => {
-    return logs.map((log) => (
-      <div key={log.id} className="group flex gap-2 py-0.5 hover:bg-slate-800/50 relative">
-        <span className="text-slate-500 shrink-0">{formatLogTime(log.timestamp)}</span>
-        <span className={`shrink-0 uppercase w-12 ${getLevelColor(log.level)}`}>
-          [{log.level}]
-        </span>
-        {log.tool !== undefined ? (
-          <span className="text-blue-400 shrink-0">[{log.tool}]</span>
-        ) : null}
-        <span className="text-slate-200 break-all whitespace-pre-wrap flex-1">{log.message}</span>
-        <button
-          type="button"
-          onClick={() => {
-            copyLogEntry(log.message, log.id);
-          }}
-          className="opacity-0 group-hover:opacity-100 transition-opacity text-slate-400 hover:text-slate-200 p-0.5"
-          title="Copy log entry"
-        >
-          {copiedLogId === log.id ? (
-            <CheckCircle2 className="h-3.5 w-3.5 text-green-400" />
-          ) : (
-            <Copy className="h-3.5 w-3.5" />
-          )}
-        </button>
-      </div>
-    ));
-  }, [logs, copiedLogId, copyLogEntry]);
-
-  return (
-    <Card className="mb-6">
-      <div className="mb-4 flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <Terminal className="h-5 w-5 text-slate-600 dark:text-slate-400" />
-          <h3 className="text-lg font-semibold text-slate-900 dark:text-slate-100">Execution Logs</h3>
-          {isActive ? (
-            <span className="flex items-center gap-1.5 rounded-full bg-green-100 px-2 py-0.5 text-xs text-green-700 dark:bg-green-900/50 dark:text-green-300">
-              <span className="h-2 w-2 rounded-full bg-green-500 animate-pulse" />
-              Live
-            </span>
-          ) : null}
-        </div>
-        <div className="flex items-center gap-3">
-          <span className="text-sm text-slate-500 dark:text-slate-400">
-            {logs.length} log{logs.length !== 1 ? 's' : ''}
-          </span>
-          <label className="flex items-center gap-2 text-sm text-slate-600 dark:text-slate-300">
-            <input
-              type="checkbox"
-              checked={autoScroll}
-              onChange={(e): void => {
-                setAutoScroll(e.target.checked);
-              }}
-              className="rounded border-slate-300 text-blue-600 focus:ring-blue-500 dark:border-slate-600 dark:bg-slate-700"
-            />
-            Auto-scroll
-          </label>
-        </div>
-      </div>
-
-      <div className="relative">
-        <div
-          ref={logsContainerRef}
-          className="rounded-lg bg-slate-900 p-4 font-mono text-sm overflow-y-auto"
-          style={{ height: `${String(logsHeight)}px` }} // eslint-disable-line no-restricted-syntax -- dynamic height value for resizable logs
-        >
-        {logsLoading ? (
-          <div className="flex items-center gap-2 text-slate-400">
-            <Loader2 className="h-4 w-4 animate-spin" />
-            Loading logs...
-          </div>
-        ) : logsError !== null ? (
-          <div className="text-red-400">Error: {logsError}</div>
-        ) : logs.length === 0 ? (
-          <div className="text-slate-500">No logs yet...</div>
-        ) : (
-          <>
-            {logsElements}
-            <div ref={logsEndRef} />
-          </>
-        )}
-        </div>
-        <div
-          ref={resizeHandleRef}
-          onMouseDown={handleMouseDown}
-          className={`absolute bottom-0 left-0 right-0 h-2 cursor-ns-resize border-t-2 border-transparent hover:border-slate-500 dark:hover:border-slate-400 transition-colors ${isResizing ? 'border-slate-500 dark:border-slate-400' : ''}`}
-          title="Drag to resize"
-        />
-      </div>
-    </Card>
-  );
-}, (prevProps, nextProps) => {
-  // Only re-render if taskId or isActive changed
-  // isActive changes update the "Live" badge, so we allow re-render for it
-  return prevProps.taskId === nextProps.taskId && prevProps.isActive === nextProps.isActive;
-});
